pyproject.toml
[project]
name = "dataviz"
version = "0.1.0"
description = ""
authors = [
    {name = "yassineba13",email = "yassine.babdallah0@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.42.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "plotly (>=6.0.0,<7.0.0)",
    "anthropic (>=0.45.2,<0.46.0)",
    "numpy (>=2.2.2,<3.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "ruff (>=0.9.5,<0.10.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "sphinx (>=8.1.3,<9.0.0)",
    "sphinx-rtd-theme (>=3.0.2,<4.0.0)"
]

[tool.poetry]
packages = [{include = "dataviz", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
